fix: files names and logo center issue
